Added AllCommands singleton class so we can load commands once and share with the version command.
