+++ src/Helper.ts
@@ -9,14 +9,12 @@ import { CommandsMap } from './command';
  * other tasks.
  */
 export default class implements Helper {
-	constructor(commandHelper: CommandHelper, yargs: Yargs, context: any, commandsMap: CommandsMap) {
+	constructor(commandHelper: CommandHelper, yargs: Yargs, context: any) {
 		this.command = commandHelper;
 		this.yargs = yargs;
 		this.context = context;
-		this.commandsMap = commandsMap;
 	};
 	command: CommandHelper;
 	yargs: Yargs;
 	context: any;
-	commandsMap: CommandsMap;
 }
